Splash working again, chipping away refactor
